+++ Programa/src/is/hotelzargo/negocio/appservices/RoomAppServicesImp.java
@@ -10,11 +10,10 @@ public class RoomAppServicesImp implements RoomAppServices {
 
 	@Override
 	public void addRoom(RoomTransfer t) throws RoomAppServicesException {
-		// TODO crear habitacion
+		//Crear habitacion
 		
 		DAOFactory fac = DAOFactory.getInstance();
-		RoomDAO dao = fac.getRoomDAO();
-		
+		RoomDAO dao = fac.getRoomDAO();		
 		
 		checkData (t);
 		
@@ -26,7 +25,7 @@ public class RoomAppServicesImp implements RoomAppServices {
 				throw new RoomAppServicesException("La habitacion ya existe");
 		}	
 		catch (RoomIntegrationException e) {
-			e.getMessage();
+			throw new RoomAppServicesException (e.getMessage());
 		}
 }
 		
@@ -47,13 +46,18 @@ public class RoomAppServicesImp implements RoomAppServices {
 	
 	@Override
 	public void delRoom(int id) throws RoomAppServicesException {
-		// TODO borrar habitacion
+		//Borrar habitacion
 		
 		DAOFactory fac = DAOFactory.getInstance();
 		RoomDAO dao = fac.getRoomDAO();
 		
 		try {
+			if (dao.searchRoomByID(id)){
 				dao.deleteRoom(id);
+			}
+			else{
+				throw new RoomAppServicesException("El ID de la habitación a eliminar no existe");
+			}
 				
 		} catch (RoomIntegrationException e) {
 			throw new RoomAppServicesException (e.getMessage());
@@ -63,7 +67,7 @@ public class RoomAppServicesImp implements RoomAppServices {
 
 	@Override
 	public void listRoom() throws RoomAppServicesException {
-		// TODO listar habitaciones
+		//Listar habitaciones
 		
 		DAOFactory fac = DAOFactory.getInstance();
 		RoomDAO dao = fac.getRoomDAO();
@@ -71,22 +75,27 @@ public class RoomAppServicesImp implements RoomAppServices {
 		try {
 			dao.listRoom();
 		} catch (RoomIntegrationException e) {
-			e.printStackTrace();
+			throw new RoomAppServicesException (e.getMessage());
 		}
 		
 	}
 
 	@Override
 	public void modRoom(RoomTransfer t) throws RoomAppServicesException {
-		// TODO modificar habitaciones
+		//Modificar habitaciones
 		
 		DAOFactory fac = DAOFactory.getInstance();
 		RoomDAO dao = fac.getRoomDAO();
 		
 		try {
-			dao.updateRoom(t);
+			if (dao.searchRoomByID(t.getId())){
+				dao.updateRoom(t);
+			}
+			else{
+				throw new RoomAppServicesException ("la habitacion a modificar no existe");
+			}
 		} catch (RoomIntegrationException e) {
-			e.printStackTrace();
+			throw new RoomAppServicesException (e.getMessage());
 		}
 		
 	}
